Refactor: Moving raster code to namespace.
