+++ NOpenType/NOpenType.csproj
@@ -39,16 +39,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Bezier.cs" />
     <Compile Include="Bounds.cs" />
     <Compile Include="CharacterMap.cs" />
     <Compile Include="Glyph.cs" />
     <Compile Include="Interpreter.cs" />
     <Compile Include="IO\ByteOrderSwappingBinaryReader.cs" />
-    <Compile Include="Line.cs" />
-    <Compile Include="Raster.cs" />
-    <Compile Include="Rasterizer.cs" />
-    <Compile Include="Segment.cs" />
     <Compile Include="Tables\BoundsReader.cs" />
     <Compile Include="Tables\CmapReader.cs" />
     <Compile Include="Tables\Glyf.cs" />
@@ -65,6 +60,11 @@
     <Compile Include="GlyphPathBuilderBase.cs" />
     <Compile Include="PathBuilderAdapter.cs" />
     <Compile Include="IGlyphRasterizer.cs" />
+    <Compile Include="Rasterizer\Line.cs" />
+    <Compile Include="Rasterizer\Raster.cs" />
+    <Compile Include="Rasterizer\Segment.cs" />
+    <Compile Include="Rasterizer\Bezier.cs" />
+    <Compile Include="Rasterizer\Rasterizer.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
@@ -74,4 +74,7 @@
   <Target Name="AfterBuild">
   </Target>
   -->
+  <ItemGroup>
+    <Folder Include="Rasterizer\" />
+  </ItemGroup>
 </Project>
